requirements.txt
neo4j==5.12.0
pinecone-client==2.2.4
openai==1.0.0
python-dotenv==1.0.0
pyyaml==6.0.1
pandas==2.1.1
numpy==1.24.3
tqdm==4.66.1